Subindo a classe Unidade e fazendo ajustes

+++ src/main/java/org/igesta/dto/GestorRequestDTO.java
@@ -14,7 +14,7 @@ import org.igesta.validation.OnCreate;
 public class GestorRequestDTO {
     @Column(name = "id_empresa")
     @Schema(description = "ID único da empresa que ogestor trabalha", example = "1234")
-    private Integer idEmpresa;
+    private Long idEmpresa;
 
     @NotNull(groups = OnCreate.class)
     @Pattern(regexp = ".*\\S.*\\S.*", message = "O nome não pode estar em branco!")
